requirements.txt
pycryptodome>=3.19.0
pillow>=10.0.0